splitted Game.tsx to GameRound1.tsx and GameRound2.tsx   still under development
