+++ src/ImageGrid.css
@@ -1,0 +1,16 @@
+.image-grid {
+    display: flex;
+    flex-wrap: wrap;
+}
+  
+.image-item {
+    flex: 1 1 calc(50% - 10px); /* Two columns with some spacing */
+    margin: 5px;
+    text-align: center; /* Center the text below the image */
+}
+  
+.image-item img {
+    max-width: 100%;
+    height: auto;
+    margin-bottom: 2px;
+}
