+++ src/components/Game/GameRoundProps.ts
@@ -1,0 +1,6 @@
+import { GameRoundStatus } from "../../utils/dataBank.ts";
+
+export interface GameRoundProps {
+    currentRoundStatus: GameRoundStatus;
+    setCurrentRoundStatus: React.Dispatch<React.SetStateAction<GameRoundStatus>>;
+}
